+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import Header from '@/components/Header/header';
+import Header from '@/components/header/header';
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import './globals.css';
